# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "DraftPreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          }
        },
        "title": "DraftPreviewRequest",
        "type": "object"
      },
      "DraftPreviewResponse": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "can_create_draft": {
            "default": false,
            "title": "Can Create Draft",
            "type": "boolean"
          },
          "cc": {
            "items": {
              "type": "string"
            },
            "title": "Cc",
            "type": "array"
          },
          "effect": {
            "default": "Approving creates a draft in the mailbox for a person to review and send. ResolveFlow never sends a message itself.",
            "title": "Effect",
            "type": "string"
          },
          "from_address": {
            "title": "From Address",
            "type": "string"
          },
          "in_reply_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Reply To"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "to": {
            "items": {
              "type": "string"
            },
            "title": "To",
            "type": "array"
          }
        },
        "required": [
          "from_address",
          "to",
          "cc",
          "subject",
          "body"
        ],
        "title": "DraftPreviewResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderDraftView": {
        "additionalProperties": false,
        "properties": {
          "failure_code": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "provider_draft_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Draft Id"
          },
          "status": {
            "enum": [
              "PENDING",
              "CREATED",
              "FAILED",
              "REFUSED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ProviderDraftView",
        "type": "object"
      },
...
      "RefusalResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Code",
            "type": "string"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "RefusalResponse",
        "type": "object"
      },
      "ReviewOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "title": "Code",
            "type": "string"
          },
          "decision": {
            "enum": [
              "EDIT",
              "APPROVE",
              "REJECT",
              "REROUTE",
              "ASSIGN",
              "RESOLVE"
            ],
            "title": "Decision",
            "type": "string"
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderDraftView"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "ticket_version": {
            "title": "Ticket Version",
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "decision",
          "code",
          "message",
          "ticket_version"
        ],
        "title": "ReviewOutcomeResponse",
        "type": "object"
      },
      "ReviewRequestBody": {
        "additionalProperties": false,
        "properties": {
          "assignee_membership_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Membership Id"
          },
          "body": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "decision": {
            "enum": [
              "EDIT",
              "APPROVE",
              "REJECT",
              "REROUTE",
              "ASSIGN",
              "RESOLVE"
            ],
            "title": "Decision",
            "type": "string"
          },
          "department_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "expected_version": {
            "minimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "queue_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Id"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "decision",
          "expected_version",
          "idempotency_key"
        ],
        "title": "ReviewRequestBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Private service contract for grounded support-ticket triage.",
    "title": "ResolveFlow AI Internal API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tickets/{ticket_id}/draft-preview": {
      "post": {
        "operationId": "draft_preview_v1_tickets__ticket_id__draft_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOutcomeResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Draft Preview",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/tickets/{ticket_id}/review": {
      "post": {
        "operationId": "review_v1_tickets__ticket_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOutcomeResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Review",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}